Move webhook trigger under app
